Use single Component with grid for ArticlePreview, Remove ArticlePreviewDensedLayout & ArticlePreviewStandardLayout
