--- src/article/preview/ArticlePreviewDensedLayout.module.scss
@@ -1,124 +1,0 @@
-@import 'breakpoints';
-
-.root {
-    width: 100%;
-    padding: var(--lotta-spacing);
-    margin-bottom: var(--lotta-spacing);
-    border-radius: var(--lotta-border-radius);
-    box-shadow: 1px 1px 2px rgba(var(--lotta-text-color), 0.2);
-    display: flex;
-    position: relative;
-    min-height: 40px; // TODO: Why 40? Make this variable
-
-    &.isEmbedded {
-        box-shadow: initial;
-    }
-
-    &:hover {
-        .editButton {
-            border: 0;
-            display: flex;
-            color: rgb(var(--lotta-primary-color));
-        }
-    }
-
-    .boxContent {
-        padding: 0 var(--lotta-spacing) !important;
-        flex: 1;
-    }
-
-    .previewText {
-        font-size: 0.85em;
-        padding-bottom: calc(0.5 * var(--lotta-spacing));
-        @media screen and (max-width: $max-sm) {
-            line-height: 1.1;
-        }
-    }
-
-    .meta {
-        flex: 0 1 auto;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: space-around;
-        & > span {
-            text-align: center;
-        }
-    }
-
-    .editSection {
-        flex: 0 1 auto;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: space-between;
-        margin-left: calc(2 * var(--lotta-spacing));
-    }
-
-    .subtitle {
-        text-transform: uppercase;
-        font-size: 0.85rem;
-        margin-bottom: 0;
-        @media screen and (max-width: $max-sm) {
-            line-height: 1.3;
-        }
-        @media screen and (max-width: $max-xs) {
-            display: flex;
-            line-height: 1.15;
-        }
-    }
-
-    .editButton {
-        float: right;
-        color: initial;
-        background: transparent;
-        transition: opacity ease-in 250ms;
-        min-height: 25px;
-        width: 25px;
-        height: 25px;
-        svg {
-            font-size: 15px !important;
-        }
-    }
-
-    .pinButton {
-        min-height: 25px;
-        width: 25px;
-        height: 25px;
-        float: right;
-        color: inherit;
-        svg {
-            font-size: 15px !important;
-        }
-        &.active {
-            color: inherit;
-        }
-    }
-
-    .previewImageWrapper {
-        flex: 0 0 20%;
-        position: relative;
-        .articlePreviewImage {
-            width: 100%;
-            background: transparent 50% 50% / cover no-repeat;
-        }
-    }
-
-    .articleTitle {
-        font-family: var(--lotta-title-font-family);
-        word-break: break-word;
-        hyphens: auto;
-        @media screen and (max-width: $max-sm) {
-            font-size: 1em;
-            font-weight: bolder;
-            line-height: 1;
-        }
-    }
-
-    .previewTextLimitedHeight {
-        overflow: hidden;
-        line-clamp: 2;
-        box-orient: vertical;
-        display: -webkit-box;
-    }
-}
