--- src/article/preview/ArticlePreviewStandardLayout.module.scss
@@ -1,175 +1,0 @@
-@import 'breakpoints';
-
-.root {
-    position: relative;
-    background-color: rgb(var(--lotta-box-background-color));
-    padding: var(--lotta-spacing);
-    margin-bottom: var(--lotta-spacing);
-    border-radius: var(--lotta-border-radius);
-    box-shadow: 1px 1px 2px rgba(var(--lotta-text-color), 0.2);
-
-    &.narrow {
-        padding: 0 0 var(--lotta-spacing) 0;
-    }
-
-    &.isEmbedded {
-        box-shadow: initial;
-    }
-    &:hover {
-        .editButton {
-            color: rgb(var(--lotta-primary-color));
-        }
-    }
-    .containerGrid {
-        display: flex;
-        flex-wrap: wrap;
-        @media (min-width: $min-sm) {
-            flex-wrap: nowrap;
-        }
-        @media print {
-            flex-wrap: nowrap;
-        }
-    }
-    &.narrow .containerGrid {
-        flex-wrap: wrap;
-    }
-
-    .previewImage {
-        width: 100%;
-    }
-    &.narrow .previewImage {
-        height: auto;
-    }
-
-    .mainSection {
-        padding-left: calc(1.5 * var(--lotta-spacing));
-        padding-right: var(--lotta-spacing);
-        flex: 1;
-        @media screen and (max-width: $max-xs) {
-            border: 0;
-            padding-right: 40px; // Make place for button + margin
-            width: 100%;
-        }
-    }
-
-    .metaSection {
-        display: flex;
-        .dateGridItem {
-            display: flex;
-            align-items: baseline;
-
-            .date {
-                padding-top: var(--lotta-spacing);
-                margin-right: calc(2 * var(--lotta-spacing));
-                @media screen and (max-width: $max-xs) {
-                    padding: calc(0.5 * var(--lotta-spacing));
-                }
-            }
-        }
-        .authorAvatarsList {
-            flex-grow: 1;
-            flex-shrink: 0;
-            width: 10em;
-        }
-    }
-
-    &.narrow .mainSection {
-        max-width: calc(100% - 40px);
-    }
-
-    .editSection {
-        @media screen and (max-width: $max-xs) {
-            position: absolute;
-            min-width: 40px; // button + margin => 32+8
-            bottom: 0;
-            right: 0;
-        }
-        @media print {
-            display: none;
-        }
-    }
-
-    .imageSection {
-        width: 100%;
-        flex-shrink: 0;
-        @media (min-width: $min-sm) {
-            width: 30%;
-        }
-        @media print {
-            width: 30%;
-        }
-    }
-    &.narrow .imageSection {
-        width: 100%;
-    }
-
-    .title {
-        padding: var(--lotta-spacing) 0;
-        font-family: var(--lotta-title-font-family);
-        font-size: 1.4rem;
-        word-break: break-word;
-        hyphens: auto;
-        @media screen and (max-width: $max-sm) {
-            font-size: 1.2rem;
-            line-height: 1.05;
-        }
-    }
-
-    .previewSection {
-        margin-bottom: var(--lotta-spacing);
-        color: rgb(var(--lotta-label-text-color));
-        overflow: hidden;
-        text-overflow: ellipsis;
-        min-height: 3.2em;
-        display: flex;
-        @media screen and (max-width: $max-xs) {
-            padding: calc(0.5 * var(--lotta-spacing));
-        }
-        @media screen and (max-width: $max-sm) {
-            display: flex !important;
-            line-height: 1.5;
-            span:last-child {
-                text-align: right;
-            }
-        }
-        span {
-            @media screen and (max-width: $max-md) {
-                display: initial;
-                width: auto;
-            }
-        }
-    }
-    &.narrow .previewSection {
-        display: block;
-        span {
-            @media screen and (max-width: $max-md) {
-                display: block;
-                width: 100%;
-            }
-        }
-    }
-
-    .buttonSection {
-        text-align: right;
-        padding-top: var(--lotta-spacing);
-        @media screen and (max-width: $max-xs) {
-            padding: calc(0.5 * var(--lotta-spacing));
-        }
-    }
-
-    .editButton {
-        color: rgb(var(--lotta-highlight-color));
-    }
-
-    .pinButton {
-        &.active {
-            svg {
-                color: rgb(var(--lotta-primary-color));
-            }
-        }
-    }
-
-    .link {
-        width: 100%;
-    }
-}
