+++ src/article/preview/index.ts
@@ -1,3 +1,1 @@
 export * from './ArticlePreview';
-export * from './ArticlePreviewDensedLayout';
-export * from './ArticlePreviewDensedLayout';
